Continue untangling includes, add filesys.c
