+++ src/groups.h
@@ -4,7 +4,6 @@
 #include "messages.h"
 
 #include "av/audio.h"
-// #include "ui/edit.h"
 
 typedef struct edit_change EDIT_CHANGE;
 
@@ -21,7 +20,7 @@ typedef struct group_peer {
     uint32_t id;
     uint32_t name_color;
     size_t name_length;
-    uint8_t name[0];
+    uint8_t name[];
 } GROUP_PEER;
 
 typedef struct groupchat {
