+++ src/ui/contextmenu.c
@@ -3,6 +3,9 @@
 #include "draw.h"
 #include "../theme.h"
 
+// FIXME: Required for UNUSED()
+#include "../main.h"
+
 static CONTEXTMENU context_menu;
 
 #define CONTEXT_WIDTH (SCALE(120))
