+++ src/ui/dropdown.c
@@ -4,6 +4,9 @@
 
 #include "../theme.h"
 
+// FIXME: Required for UNUSED()
+#include "../main.h"
+
 static DROPDOWN *active_dropdown;
 static int       active_x, active_y, active_width, active_height;
