+++ src/ui/switch.c
@@ -3,6 +3,9 @@
 #include "draw.h"
 #include "tooltip.h"
 
+// FIXME: Required for UNUSED()
+#include "../main.h"
+
 static void calculate_pos_and_width(UISWITCH *s, int *x, int *w) {
     int old_w = *w;
